refactor: palette class OOP refactoring
Closes #4
